diy/requirements.txt
mysql-connector-python==8.0.30
matplotlib==3.5.3
extra-streamlit-components==0.1.55
python-linkedin-v2==0.9.4
pydeck==0.7.1
pandas==1.3.5
numpy==1.21.6
websocket-client===1.3.3
simple-websocket-server===0.4.2
rel===0.4.7
websockets===10.3
